#16 move model to api package
There where some TypeScript issues with importing runtypes definitions
from another package in connection with rest.ts
I was not able to resolve them, but a separate model package is
atm not really needed.
