+++ packages/api/package.json
@@ -26,7 +26,6 @@
     "typescript": "^3.1.6"
   },
   "dependencies": {
-    "@open-gtd/model": "1.0.0",
     "rest-ts-core": "0.3.2",
     "runtypes": "^2.1.6"
   }
