--- packages/model/.gitignore
@@ -1,2 +1,0 @@
-lib/
-node_modules/
