--- packages/model/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@open-gtd/model",
-  "version": "1.0.0",
-  "private": true,
-  "description": "",
-  "main": "lib/index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\"",
-    "build": "tsc",
-    "start": "tsc --watch --preserveWatchOutput",
-    "lint:tslint": "tslint --project . --config ../../tslint.prod.json",
-    "fix:tslint": "tslint --project . --fix",
-    "lint:prettier": "prettier-check src/**/*",
-    "fix:prettier": "prettier --write src/**/*",
-    "lint": "npm-run-all lint:**",
-    "fix": "npm-run-all fix:**",
-    "verify": "npm-run-all build lint test"
-  },
-  "author": "",
-  "license": "BSD-3-Clause",
-  "devDependencies": {
-    "npm-run-all": "^4.1.3",
-    "prettier": "^1.15.2",
-    "prettier-check": "^2.0.0",
-    "tslint": "^5.11.0",
-    "typescript": "^3.1.6"
-  },
-  "dependencies": {
-    "runtypes": "^2.1.6"
-  }
-}
